// src/router/router.types.ts
export enum RouterNameEnum {
    Main = 'Main',
    AuthLogin = 'AuthLogin',
    AuthRecover = 'AuthRecover',

    Salons = 'Salons',
    SalonsPage = 'SalonsPage',
    LoyaltyCertPage = 'LoyaltyCertPage',

    Reports = 'Reports',
    ReportSalesByService = 'ReportSalesByService',

    Store = 'Store',
    StoreDetail = 'StoreDetail',

    Stores = 'Stores',
    StoresPage = 'StoresPage',
    StoresProducts = 'StoresProducts',
    StoresProductsPage = 'StoresProductsPage',

    StoresProductsArchived = 'StoresProductsArchived',
    StoresProductsPageArchived = 'StoresProductsPageArchived',


    StoresProductsCreate = 'StoresProductsCreate',
    StoresProductPage = 'StoresProductPage',
    StoresProductsMassEdit = 'StoresProductsMassEdit',
    StoresProductsMassEditPage = 'StoresProductsMassEditPage',
    StoreOperations = 'StoreOperations',
    StoreSettings = 'StoreSettings',

    IndexStore = 'IndexStore',
    CreateStore = 'CreateStore',

    Clients = 'Clients',
    ClientsPage = 'ClientsPage',

    Settings = 'Settings',
    Loyalty = 'Loyalty',
    SalonDetail = 'SalonDetail',
    SalonCreation = 'SalonCreation',
    SalonDetailCommon = 'SalonDetailCommon',
    SalonDetailFinance = 'SalonDetailFinance',
    SalonDetailReports = 'SalonDetailReports',
    SalonDetailClients = 'SalonDetailClients',
    SalonDetailClientsPage = 'SalonDetailClientsPage',

    SalonDetailWorkers = 'SalonDetailWorkers',
    SalonDetailPhoto = 'SalonDetailPhoto',
    SalonDetailDocs = 'SalonDetailDocs',
    SalonDetailAllow = 'SalonDetailAllow',
    ClientDetail = 'ClientDetail',
    ClientCreation = 'ClientCreation',
    ClientDetailCommon = 'ClientDetailCommon',
    ClientDetailHistory = 'ClientDetailHistory',
    ClientDetailNotes = 'ClientDetailNotes',
    ClientDetailTasks = 'ClientDetailTasks',
    ClientDetailDocuments = 'ClientDetailDocuments',
    ClientDetailPhoto = 'ClientDetailPhoto',
    ClientDetailContraindications = 'ClientDetailContraindications',
    ClientDetailLoyalty = 'ClientDetailLoyalty',
    Finance = 'Finance',
    AccountsAndCashboxes = 'AccountsAndCashboxes',
    Counterparties = 'Counterparties',
    PaymentArticles = 'PaymentArticles',
    FinanceDocuments = 'FinanceDocuments',
    FinanceOperations = 'FinanceOperations',
    Services = 'Services',
    LoyaltyCert = 'LoyaltyCert',
    LoyaltyCertType = 'LoyaltyCertType',
    LoyaltySubscription = 'LoyaltySubscription',
    LoyaltySubscriptionType = 'LoyaltySubscriptionType',
    LoyaltyBonus = 'LoyaltyBonus',
    LoyaltyBonusType = 'LoyaltyBonusType',
    LoyaltyStock = 'LoyaltyStock',
    ServicesCreation = 'ServicesCreation',
    ServiceUpdate = 'ServiceUpdate',
    ServicesCreationCosmetologists = 'ServicesCreationCosmetologists',
    ServicesCreationConsumables = 'ServicesCreationConsumables',
    ServicesCreationDocuments = 'ServicesCreationDocuments',
    ServicesCreationAnamnez = 'ServicesCreationAnamnez',
    ServicesCreationAlgoritm = 'ServicesCreationAlgoritm',
    ServicesCategory = 'ServicesCategory',
    ServicesCategoryCreation = 'ServicesCategoryCreation',
    ServiceDetail = 'ServiceDetail',
    ServiceDetailWorkers = 'ServiceDetailWorkers',
    ServiceDetailProducts = 'ServiceDetailProducts',
    ServiceDetailDocuments = 'ServiceDetailDocuments',
    ServiceDetailAnamnez = 'ServiceDetailAnamnez',
    Visits = 'Visits',
    VisitsDetail = 'VisitDetail'
}